requirements.txt
allure-pytest==2.13.2
allure-python-commons==2.13.2
click==8.1.6
distlib==0.3.7
iniconfig==2.0.0
outcome==1.2.0
packaging==23.1
pluggy==1.0.0
pre-commit==3.3.2
pytest==7.3.1
requests==2.31.0
selenium==4.15.2
urllib3==1.26.15
